+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+### Changed
+- Load upper player limit from options
+
 ## [v0.3.1] - 2020-02-27
 ### Fixed
 - Wrong calculation of the card value